requirements.txt
crewai~=0.159.0
langchain-openai~=0.2.14
python-dotenv~=1.1.1